+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - No new features!
 ### Changed
 - The xml attribute "encoding" in the generated string resource files is now lowercase "utf-8"
+- Migrate to PoEditor API v2
 ### Deprecated
 - No deprecated features!
 ### Removed